Added first tests and moved function
Added first tests of the "needSpace" function
Moved "needSpace" function in seperate file
