+++ .gitignore
@@ -60,4 +60,3 @@ typings/
 
 # dotenv environment variables file
 .env
-
